# specs/openapi.yaml
openapi: 3.1.0
info:
  title: VirtualRadio API
  version: 1.0.0
  description: |
    API REST de **VirtualRadio**: plataforma que genera de forma automatizada
    programas de radio satíricos (estilo WCTR) para videojuegos de simulación
    (Farming Simulator 25, Euro Truck Simulator 2, etc.).

    El backend orquesta una biblioteca musical local, un universo narrativo
    compartido (noticias, comerciales, personajes y memoria narrativa) y un
    pipeline de generación por agentes que produce episodios de audio listos
    para reproducir.

    ## Autenticación
    Autenticación mediante **JWT** (Flask-JWT-Extended). `POST /auth/login`
    emite un *access token* y un *refresh token*; `POST /auth/refresh` renueva
    el *access token* a partir de un *refresh token* válido. Todos los endpoints
    salvo los de `auth` y `health` requieren `Authorization: Bearer <access_token>`.

    ## Autorización (RBAC)
    Modelo deliberadamente simple: un único rol de sistema `USER` y un único
    scope efectivo `own`. Cada consulta se filtra por `owner_id = current_user.id`.
    El `owner_id` se toma **siempre del token**, nunca del body. El acceso a un
    recurso de otro usuario devuelve **404** (no 403) para no revelar su existencia.

    ## Contrato de respuestas
    - Éxito: `{ "data": ..., "meta": {...} }`
    - Error: `{ "error": { "code", "message", "details" } }`
  contact:
    name: Equipo de Plataforma / Backend
  license:
    name: Propietario
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /news:
    get:
      tags: [News]
      summary: Listar noticias
      description: 'Permiso: `news:read:own`.'
      operationId: listNews
      parameters:
        - name: is_active
          in: query
          required: false
          description: Filtrar por estado activo
          schema:
            type: boolean
        - name: category
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/NewsCategory'
      responses:
        '200':
          description: Lista de noticias
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsItemList'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [News]
      summary: Crear noticia
      description: 'Permiso: `news:create:own`.'
      operationId: createNews
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsItemInput'
      responses:
        '201':
          description: Noticia creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsItemResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /news/suggest:
    post:
      tags: [News]
      summary: Sugerir noticia por IA
      description: 'Permiso: `news:suggest:own`. Fallback procedural si el LLM falla.'
      operationId: suggestNews
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SuggestRequest'
      responses:
        '200':
          description: Sugerencia de noticia (no persistida)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsItemResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /news/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [News]
      summary: Obtener noticia
      description: 'Permiso: `news:read:own`.'
      operationId: getNews
      responses:
        '200':
          description: Noticia
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsItemResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [News]
      summary: Actualizar noticia
      description: 'Permiso: `news:update:own`.'
      operationId: updateNews
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsItemInput'
      responses:
        '200':
          description: Noticia actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsItemResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [News]
      summary: Eliminar noticia
      description: 'Permiso: `news:delete:own`.'
      operationId: deleteNews
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------------------
  # Brands
  # ---------------------------------------------------------------------------
  /brands:
    get:
      tags: [Brands]
      summary: Listar marcas
      description: 'Permiso: `brand:read:own`.'
      operationId: listBrands
      responses:
        '200':
          description: Lista de marcas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandList'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Brands]
      summary: Crear marca
      description: 'Permiso: `brand:create:own`.'
      operationId: createBrand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrandInput'
      responses:
        '201':
          description: Marca creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /brands/suggest:
    post:
      tags: [Brands]
      summary: Sugerir marca por IA
      description: 'Permiso: `brand:suggest:own`. Fallback procedural si el LLM falla.'
      operationId: suggestBrand
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SuggestRequest'
      responses:
        '200':
          description: Sugerencia de marca (no persistida)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /brands/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Brands]
      summary: Obtener marca
      description: 'Permiso: `brand:read:own`.'
      operationId: getBrand
      responses:
        '200':
          description: Marca
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Brands]
      summary: Actualizar marca
      description: 'Permiso: `brand:update:own`.'
      operationId: updateBrand
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrandInput'
      responses:
        '200':
          description: Marca actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [Brands]
      summary: Eliminar marca
      description: 'Permiso: `brand:delete:own`. Borra en cascada sus comerciales.'
      operationId: deleteBrand
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------------------
  # Commercials
  # ---------------------------------------------------------------------------
  /commercials:
    get:
      tags: [Commercials]
      summary: Listar comerciales
      description: 'Permiso: `commercial:read:own`.'
      operationId: listCommercials
      parameters:
        - name: brand_id
          in: query
          required: false
          description: Filtrar comerciales por marca
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista de comerciales
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommercialList'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Commercials]
      summary: Crear comercial
      description: 'Permiso: `commercial:create:own`. Requiere una marca existente del usuario.'
      operationId: createCommercial
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommercialInput'
      responses:
        '201':
          description: Comercial creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommercialResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: La marca indicada no existe o pertenece a otro usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/ValidationError'
  /commercials/suggest:
    post:
      tags: [Commercials]
      summary: Sugerir comercial por IA
      description: 'Permiso: `commercial:suggest:own`. Fallback procedural si el LLM falla.'
      operationId: suggestCommercial
      requestBody:
        required: false
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/SuggestRequest'
                - type: object
                  properties:
                    brand_id:
                      type: string
                      format: uuid
                      description: Marca para la que generar el comercial
      responses:
        '200':
          description: Sugerencia de comercial (no persistida)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommercialResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /commercials/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Commercials]
      summary: Obtener comercial
      description: 'Permiso: `commercial:read:own`.'
      operationId: getCommercial
      responses:
        '200':
          description: Comercial
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommercialResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Commercials]
      summary: Actualizar comercial
      description: 'Permiso: `commercial:update:own`.'
      operationId: updateCommercial
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommercialInput'
      responses:
        '200':
          description: Comercial actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommercialResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [Commercials]
      summary: Eliminar comercial
      description: 'Permiso: `commercial:delete:own`.'
      operationId: deleteCommercial
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------------------
  # Characters
  # ---------------------------------------------------------------------------
  /characters:
    get:
      tags: [Characters]
      summary: Listar personajes
      description: 'Permiso: `character:read:own`.'
      operationId: listCharacters
      responses:
        '200':
          description: Lista de personajes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CharacterList'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Characters]
      summary: Crear personaje
      description: 'Permiso: `character:create:own`.'
      operationId: createCharacter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CharacterInput'
      responses:
        '201':
          description: Personaje creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CharacterResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /characters/suggest:
    post:
      tags: [Characters]
      summary: Sugerir personaje por IA
      description: 'Permiso: `character:suggest:own`. Fallback procedural si el LLM falla.'
      operationId: suggestCharacter
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SuggestRequest'
      responses:
        '200':
          description: Sugerencia de personaje (no persistida)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CharacterResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /characters/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Characters]
      summary: Obtener personaje
      description: 'Permiso: `character:read:own`.'
      operationId: getCharacter
      responses:
        '200':
          description: Personaje
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CharacterResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Characters]
      summary: Actualizar personaje
      description: 'Permiso: `character:update:own`.'
      operationId: updateCharacter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CharacterInput'
      responses:
        '200':
          description: Personaje actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CharacterResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [Characters]
      summary: Eliminar personaje
      description: 'Permiso: `character:delete:own`. Borra en cascada sus memorias.'
      operationId: deleteCharacter
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /characters/{id}/memories:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Characters]
      summary: Memorias del personaje
      description: |
        Devuelve la memoria narrativa del personaje (eventos pasados que
        mantienen la continuidad). Permiso: `character:read_memories:own`.
      operationId: getCharacterMemories
      responses:
        '200':
          description: Lista de memorias
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CharacterMemoryList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------------------------------------------------------------------------
  # Story Events
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Token JWT emitido por `/auth/login` o `/auth/refresh`.'
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      description: Identificador UUID del recurso
      schema:
        type: string
        format: uuid
  responses:
    NoContent:
      description: Operación realizada con éxito (sin contenido)
    Unauthorized:
      description: Token ausente, inválido o expirado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHORIZED
              message: Token inválido o ausente
              details: {}
    NotFound:
      description: |
        El recurso no existe o pertenece a otro usuario (se devuelve 404 en
        ambos casos para no revelar su existencia).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: NOT_FOUND
              message: Recurso no encontrado
              details: {}
    ValidationError:
      description: Error de validación de los datos enviados
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: VALIDATION_ERROR
              message: Los datos enviados no son válidos
              details:
                name: ["El campo es obligatorio"]
  schemas:
...
    Meta:
      type: object
      description: Metadatos de la respuesta (paginación, conteos, etc.)
      additionalProperties: true
      example: {}
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Código de error legible por máquina
              example: VALIDATION_ERROR
            message:
              type: string
              description: Descripción del error
            details:
              type: object
              description: Detalles adicionales (ej. errores por campo)
              additionalProperties: true
    Timestamps:
      type: object
      properties:
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true

    # ---- Enums ----------------------------------------------------------------
    NewsCategory:
      type: string
      description: Categoría temática de la noticia
      enum:
        - Agricultura
        - Transporte
        - Economía
        - Tecnología
        - Clima
        - Comunidad
        - Política Local
        - Sucesos Extraños
    NewsTone:
      type: string
      description: Tono narrativo de la noticia
      enum:
        - Serio
        - Sensacionalista
        - Misterioso
        - Absurdo
...
    SuggestRequest:
      type: object
      description: Pista/contexto opcional para guiar la sugerencia del LLM.
      properties:
        prompt:
          type: string
          description: Indicación libre para orientar la generación
        context:
          type: object
          description: Contexto adicional (ej. estación, tema)
          additionalProperties: true

    # ---- Stations -------------------------------------------------------------
...
    NewsItemInput:
      type: object
      required: [headline, category, tone]
      properties:
        headline:
          type: string
          maxLength: 300
        summary:
          type: string
          nullable: true
        full_script:
          type: string
          nullable: true
          description: Guion de lectura (3-5 frases)
        category:
          $ref: '#/components/schemas/NewsCategory'
        tone:
          $ref: '#/components/schemas/NewsTone'
        is_active:
          type: boolean
          default: true
        expires_at:
          type: string
          format: date-time
          nullable: true
    NewsItem:
      allOf:
        - type: object
          properties:
            id:
              type: string
              format: uuid
              readOnly: true
            owner_id:
              type: string
              format: uuid
              readOnly: true
        - $ref: '#/components/schemas/NewsItemInput'
        - $ref: '#/components/schemas/Timestamps'
    NewsItemResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/NewsItem'
        meta:
          $ref: '#/components/schemas/Meta'
    NewsItemList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/NewsItem'
        meta:
          $ref: '#/components/schemas/Meta'

    # ---- Brands ---------------------------------------------------------------
    BrandInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 150
          description: Nombre de la marca (único por usuario)
        description:
          type: string
          nullable: true
        industry:
          type: string
          maxLength: 100
          nullable: true
        slogan:
          type: string
          maxLength: 255
          nullable: true
        is_active:
          type: boolean
          default: true
    Brand:
      allOf:
        - type: object
          properties:
            id:
              type: string
              format: uuid
              readOnly: true
            owner_id:
              type: string
              format: uuid
              readOnly: true
        - $ref: '#/components/schemas/BrandInput'
        - $ref: '#/components/schemas/Timestamps'
    BrandResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Brand'
        meta:
          $ref: '#/components/schemas/Meta'
    BrandList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Brand'
        meta:
          $ref: '#/components/schemas/Meta'

    # ---- Commercials ----------------------------------------------------------
    CommercialInput:
      type: object
      required: [brand_id, title, script]
      properties:
        brand_id:
          type: string
          format: uuid
          description: Marca asociada (debe pertenecer al usuario)
        title:
          type: string
          maxLength: 200
        script:
          type: string
          description: Guion del comercial (3-5 frases)
        duration:
          type: number
          format: float
          default: 30
        campaign:
          type: string
          maxLength: 150
          nullable: true
        is_active:
          type: boolean
          default: true
    Commercial:
      allOf:
        - type: object
          properties:
            id:
              type: string
              format: uuid
              readOnly: true
            owner_id:
              type: string
              format: uuid
              readOnly: true
        - $ref: '#/components/schemas/CommercialInput'
        - $ref: '#/components/schemas/Timestamps'
    CommercialResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Commercial'
        meta:
          $ref: '#/components/schemas/Meta'
    CommercialList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Commercial'
        meta:
          $ref: '#/components/schemas/Meta'

    # ---- Characters -----------------------------------------------------------
    CharacterInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 150
          description: Nombre del personaje (único por usuario)
        role:
          type: string
          maxLength: 150
          nullable: true
          example: "Granjero jubilado"
        description:
          type: string
          nullable: true
        personality:
          type: string
          nullable: true
        station_affinity:
          type: string
          nullable: true
          description: Estaciones que escucha (lista separada por comas)
    Character:
      allOf:
        - type: object
          properties:
            id:
              type: string
              format: uuid
              readOnly: true
            owner_id:
              type: string
              format: uuid
              readOnly: true
            first_appearance:
              type: string
              format: date-time
              readOnly: true
            last_appearance:
              type: string
              format: date-time
              readOnly: true
        - $ref: '#/components/schemas/CharacterInput'
        - $ref: '#/components/schemas/Timestamps'
    CharacterResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Character'
        meta:
          $ref: '#/components/schemas/Meta'
    CharacterList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Character'
        meta:
          $ref: '#/components/schemas/Meta'
    CharacterMemory:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        owner_id:
          type: string
          format: uuid
          readOnly: true
        character_id:
          type: string
          format: uuid
        episode_id:
          type: string
          format: uuid
          nullable: true
        memory:
          type: string
          description: Hecho narrativo recordado
        importance:
          type: integer
          minimum: 1
          maximum: 10
          default: 5
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
    CharacterMemoryList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/CharacterMemory'
        meta:
          $ref: '#/components/schemas/Meta'

    # ---- Story Events ---------------------------------------------------------
...